requirements.txt
Flask
pymongo
python-dotenv
flask-login
bcrypt
Pillow
python-jose
requests